move interfaces to /contracts
