+++ src/contracts/generalInterfaces.ts
