+++ src/contracts/githubActionsInterface.ts
